user_management_portal/backend/requirements.txt
fastapi>=0.100.0,<0.101.0
uvicorn>=0.23.0,<0.24.0
sqlalchemy>=2.0.0,<2.1.0
pydantic>=2.0.0,<3.0.0
python-dotenv>=1.0.0,<1.1.0
email-validator>=2.0.0,<3.0.0
python-jose[cryptography]>=3.3.0,<3.4.0
passlib[bcrypt]>=1.7.4,<1.8.0
bcrypt>=4.0.1,<4.1.0
python-multipart>=0.0.6,<0.0.7
